# src/routes/authRoutes.ts
import express, { Request, Response } from 'express';
import { AuthenticatedRequest } from '../middleware/auth';
import * as authService from '../services/authService';

const router = express.Router();

router.post('/register', async (req: AuthenticatedRequest, res: Response) => {
  try {
    const { username, password } = req.body;

    await authService.registerUser(username, password);

    res.status(201).json({ message: 'User registered successfully' });
  } catch (error) {
    console.error(error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

router.post('/login', async (req: Request, res: Response) => {
  try {
    const { username, password } = req.body;

    const token = await authService.loginUser(username, password);

    res.json({ token });
  } catch (error) {
    console.error(error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

export default router;